Moving the ColorRgba type down to TerraFX.Numerics in the core assembly
