+++ samples/TerraFX/Vertices/IdentityVertex.cs
@@ -1,6 +1,5 @@
 // Copyright © Tanner Gooding and Contributors. Licensed under the MIT License (MIT). See License.md in the repository root for more information.
 
-using TerraFX.Graphics;
 using TerraFX.Numerics;
 
 namespace TerraFX.Samples.Graphics;
